move function to search results page and link to download button
