+++ client/src/components/data-table.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import Highlighter from 'react-highlight-words';
 import { Table, Input, Button, Space } from 'antd';
 import { SearchOutlined } from '@ant-design/icons';
-import downloadFileToClient from '../functions/downloadFileToClient';
 
 export default class DataTable extends React.Component {
   constructor(props) {
@@ -173,33 +172,8 @@ export default class DataTable extends React.Component {
     return columns;
   };
 
-  onChange = (pagination, filters, sorter, extra) => {
-    this.setState({ filteredData: extra.currentDataSource });
-  };
-
-  resultsToTSV = () => {
-    const resultsToExport = this.filteredData.map(
-      ({ key, ...keepAttrs }) => keepAttrs
-    );
-    const replacer = (key, value) => (value === null ? '' : value); //TODO: how to handle nulls?
-    const delimeter = '\t'; //TODO: can easily change to CSV
-    const fileExtension = 'tsv';
-    const header = Object.keys(resultsToExport[0]);
-    let tsv = [
-      header.join(delimeter),
-      ...resultsToExport.map((row) =>
-        header
-          .map((fieldName) => JSON.stringify(row[fieldName], replacer))
-          .join(delimeter)
-      ),
-    ]
-      .join('\r\n')
-      .replaceAll('"', '');
-
-    downloadFileToClient(
-      new Blob([tsv], { type: 'text/' + fileExtension }),
-      'report.' + fileExtension
-    ); //TODO: come up with a useful filename template
+  onTableChange = (pagination, filters, sorter, extra) => {
+    this.props.setDisplayedData(extra.currentDataSource)
   };
 
   render() {
@@ -224,7 +198,7 @@ export default class DataTable extends React.Component {
           defaultPageSize: 100,
           pageSizeOptions: [100, 250, 500],
         }}
-        onChange={this.onChange}
+        onChange={this.onTableChange}
       />
     );
   }
