+++ client/src/components/header.jsx
@@ -26,7 +26,7 @@ export default function Header(props) {
             type='primary'
             shape='round'
             icon={<DownloadOutlined />}
-            onClick={props.downloadOnClick}
+            onClick={props.onClickDownload}
           >
             TSV
           </Button>
